LayoutTests/imported/w3c/web-platform-tests/navigation-api/per-entry-events/dispose-same-document-intercept-expected.txt


FAIL dispose events when forward-pruning same-document entries assert_equals: expected 3 but got 4
